Expose `float dt_image_distance_transform()`
Make this function available to the rest of dt.
